svn merge -c 29304 https://source.sakaiproject.org/svn/citations/trunk U    citations-api/api/src/java/org/sakaiproject/citation/api/CitationCollection.java U    citations-impl/impl/src/java/org/sakaiproject/citation/impl/BaseCitationService.java svn log -r 29304 https://source.sakaiproject.org/svn/citations/trunk ------------------------------------------------------------------------ r29304 | gbhatnag@umich.edu | 2007-04-19 16:09:15 -0400 (Thu, 19 Apr 2007) | 2 lines
SAK-9547: added more keys to sort list on other than just title.  We're now sorting on title, then author, then id.  This allows for citations with the same title, or the same author, or the same title and author to be added.  Consequently, the same citation can be added (each citation's id is universally unique).

------------------------------------------------------------------------


git-svn-id: https://source.sakaiproject.org/svn/citations/branches/sakai_2-4-x@29627 66ffb92e-73f9-0310-93c1-f5514f145a0a